pyproject.toml
[tool.poetry]
name = "chained-notes"
version = "0.1.0"
description = "An application for chained notes."
authors = []
license = "MIT"
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.1"
django-environ = "^0.11.2"
djangorestframework = "^3.14.0"
psycopg = "^3.1.17"
psycopg-binary = "^3.1.17"


[tool.poetry.group.dev.dependencies]
pylint = "^3.0.3"
black = "^24.1.1"
isort = "^5.13.2"
pylint-django = "^2.5.5"
pyright = "^1.1.349"
django-types = "^0.19.1"
djangorestframework-types = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
